# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "TrustPay Test Api",
        "version": "1.0.0"
    },
    "basePath": "/apitest",
    "paths": {
...
        "/customer/register": {
            "post": {
                "tags": [
                    "Buyer Module"
                ],
                "summary": "Create New Customer",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "body",
                        "description": "Customer Details [CustomerObject]",
                        "required": true,
                        "schema": {
                            "properties": {
                                "firstname": {
                                    "description": "Customer Firstname",
                                    "type": "string"
                                },
                                "lastname": {
                                    "description": "Customer Lastname",
                                    "type": "string"
                                },
                                "phone_number": {
                                    "description": "Customer Phone Number",
                                    "type": "string"
                                },
                                "email": {
                                    "description": "Customer Email Address",
                                    "type": "string"
                                },
                                "password": {
                                    "description": "Customer Password",
                                    "type": "string"
                                },
                                "password_confirmation": {
                                    "description": "Confirm Password",
                                    "type": "string"
                                },
                                "PIN": {
                                    "description": "Customer PIN",
                                    "type": "integer"
                                },
                                "PIN_confirmation": {
                                    "description": "Confirm Customer PIN",
                                    "type": "integer"
                                },
                                "ip_address": {
                                    "description": "Customer IP Address",
                                    "type": "string"
                                },
                                "device_type": {
                                    "description": "Customer Device Type",
                                    "type": "string"
                                },
                                "device_version": {
                                    "description": "Customer Device Version",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status=success, message=customer_registered}"
                    },
                    "201": {
                        "description": "{status=error, message=messageText}"
                    },
                    "401": {
                        "description": "{status: error, message: invalid token/token expired, debug: [DebugInfoObject]}"
                    }
                }
            }
        },
        "/customer/confirmation/{token}": {
            "get": {
                "tags": [
                    "Buyer Module"
                ],
                "summary": "Customer Email Confirmation",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token from mail sent to user",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status: success, message: customer_verified}"
                    },
                    "201": {
                        "description": "{status: error, message: invalid_token}"
                    }
                }
            }
        },
        "/customer": {
            "get": {
                "tags": [
                    "Buyer Module"
                ],
                "summary": "Get Authenticated Customer Details",
                "responses": {
                    "200": {
                        "description": "{status: success, message: customer_details, data: [CustomerObject]}"
                    },
                    "201": {
                        "description": "{status: error, message: failed}"
                    },
                    "401": {
                        "description": "{status: error, message: invalid token/token expired, debug: [DebugInfoObject]}"
                    }
                }
            }
        },
        "/customer/update": {
            "put": {
                "tags": [
                    "Buyer Module"
                ],
                "summary": "Update/Edit Customer Details",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "body",
                        "description": "Customer Details [CustomerObject]",
                        "required": true,
                        "schema": {
                            "properties": {
                                "firstname": {
                                    "description": "Customer Firstname",
                                    "type": "string"
                                },
                                "lastname": {
                                    "description": "Customer Lastname",
                                    "type": "string"
                                },
                                "phone_number": {
                                    "description": "Customer Phone Number",
                                    "type": "string"
                                },
                                "email_address": {
                                    "description": "Customer Email Address",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status: success, message: customer_updated, data: [CustomerObject]}"
                    },
                    "201": {
                        "description": "{status: error, message: some parameters missing or required/failed}"
                    },
                    "401": {
                        "description": "{status: error, message: invalid token/token expired, debug: [DebugInfoObject]}"
                    }
                }
            }
        },
        "/customer/orders": {
            "get": {
                "tags": [
                    "Order Module"
                ],
                "summary": "Get Orders for Authenticated Customer",
                "responses": {
                    "200": {
                        "description": "{status: success, message: customer_orders, data: [TransactionObject + SellerObject]}"
                    },
                    "201": {
                        "description": "{status: error, message: failed"
                    },
                    "401": {
                        "description": "{status: error, message: invalid token/token expired, debug: [DebugInfoObject]}"
                    }
                }
            }
        },
        "/customer/order": {
            "post": {
                "tags": [
                    "Order Module"
                ],
                "summary": "Get Single Order Details",
                "parameters": [
                    {
                        "name": "id",
                        "in": "body",
                        "description": "Customer id",
                        "required": true,
                        "schema": {
                            "properties": {
                                "id": {
                                    "description": "Order Unique Id greater than 1",
                                    "type": "integer",
                                    "format": "int64"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status: success, message:order_details, data: [TransactionObject"
                    },
                    "201": {
                        "description": "status: error, message:Missing parameter Required / Failed"
                    },
                    "401": {
                        "description": "{status: error, message: invalid token/token expired, debug: [DebugInfoObject]}"
                    }
                }
            }
        },
        "/customer/order/disputes": {
            "get": {
                "tags": [
                    "Order Module"
                ],
                "summary": "Get All Disputes for Authenticated Customer",
                "responses": {
                    "200": {
                        "description": "{status: success, message:customer_disputes, data: [RefundRequestObject]}"
                    },
                    "201": {
                        "description": "{status: error, message: Could not Get Disputes}"
                    },
                    "401": {
                        "description": "{status: error, message: invalid token/token expired, debug: [DebugInfoObject]}"
                    }
                }
            }
        },
        "/customer/order/dispute": {
            "post": {
                "tags": [
                    "Order Module"
                ],
                "summary": "Get Single Dispute Details",
                "parameters": [
                    {
                        "name": "dispute_id",
                        "in": "body",
                        "description": "Dispute Unique Id",
                        "required": true,
                        "schema": {
                            "properties": {
                                "dispute_id": {
                                    "description": "Dispute Unique Id greater than 1",
                                    "type": "integer",
                                    "format": "int64"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status: success, message:dispute_details, data: [RefundRequestObject + TransactionObject]}"
                    },
                    "201": {
                        "description": "{status: error, message:Some Missing Parameters are Required / Failed}"
                    },
                    "401": {
                        "description": "{status: error, message: invalid token/token expired, debug: [DebugInfoObject]}"
                    }
                }
            }
        },
...
    },
...
}